fix(checker): move distributive-check classifier into solver type_queries
Addresses arch_guard lint failure on the prior head: `query_boundaries/common.rs`
must not directly match `TypeData` internals, and is already at its 1920-line
quarantine ceiling.

The classifier ("distributive Conditional whose check side is
Lazy/Application/IndexAccess/KeyOf") is a structural type-shape rule and
belongs in `tsz_solver::type_queries`. Adds
`is_distributive_conditional_with_deferred_check` to
`type_queries::classifiers` and re-exports it via the `type_queries`
module surface. The two checker eager-eval gates
(`state/type_analysis/computed/type_alias_variable_alias.rs` and
`types/type_checking/type_alias_checking.rs`) call it directly without
growing `query_boundaries/common.rs`.

No behavior change relative to the prior head — the helper is moved, not
modified.

Refs PR review comment from Reviewer on #12077.
