+++ crates/tsz-checker/src/state/type_analysis/computed/type_alias_variable_alias.rs
@@ -200,7 +200,7 @@ impl<'a> CheckerState<'a> {
                         && !crate::query_boundaries::common::contains_type_parameters(
                             db, alias_type,
                         )
-                        && !crate::query_boundaries::common::conditional_check_defers_into_alias(
+                        && !tsz_solver::type_queries::is_distributive_conditional_with_deferred_check(
                             db, alias_type,
                         )
                     {
