+++ crates/tsz-solver/src/type_queries/mod.rs
@@ -58,8 +58,8 @@ pub use classifiers::{
     classify_for_constructor_access, classify_for_excess_properties, classify_for_interface_merge,
     get_application_lazy_def_id, get_conditional_type_id, get_keyof_type, get_lazy_def_id,
     get_mapped_type_id, get_type_query_symbol_ref, indexed_access_self_keyof,
-    is_deferred_lazy_or_indexed_access, is_deferred_type_operation, is_indexed_access,
-    is_only_false_or_never,
+    is_deferred_lazy_or_indexed_access, is_deferred_type_operation,
+    is_distributive_conditional_with_deferred_check, is_indexed_access, is_only_false_or_never,
 };
 // `get_def_id` is an alias for `get_lazy_def_id` (identical semantics).
 pub use classifiers::get_lazy_def_id as get_def_id;
